// apps/orbit/lib/types/overview.ts
import type { DateRange } from 'react-day-picker';

export type PeriodValue = 'previous-period' | 'last-year' | 'no-comparison';

export type KpiEntry = {
  title: string;
  percentage: number;
  current: number;
  allowed: number;
  unit?: string;
};

export type KpiEntryExtended = Omit<KpiEntry, 'current' | 'allowed' | 'unit'> & {
  value: string;
  color: string;
};

export type { DateRange };